perfect except for split

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: monoue <marvin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/22 13:50:34 by monoue            #+#    #+#             */
-/*   Updated: 2020/07/02 11:38:09 by monoue           ###   ########.fr       */
+/*   Updated: 2020/07/06 15:45:11 by monoue           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int len;
+	size_t	len;
 
 	len = 0;
 	while (*str != '\0')
